my_api_requirements.txt
Django==1.8.2
django-cors-headers==1.1.0
django-oauth-toolkit==0.8.1
djangorestframework==3.1.3
Markdown==2.6.2
django-filter==0.10.0
django-extensions==1.5.5
python-social-auth==0.2.10
